requirements.txt
# requirements.txt

africastalking==1.2.0
Django==5.0.7
djangorestframework==3.15.2
gunicorn==22.0.0
# ... and other libraries
